canbus added

+++ include/PINOUT.h
@@ -45,10 +45,10 @@
 #define VNQ5050KTRE_LEFT_STAT_DIS_PIN           18
 #define VNQ5050KTRE_LEFT_STAT_DIS_PORT          hetPORT1
 
-uint16 VNQ5050KTRE_LEFT_1STAT1_VAL;             //ADIN[11]
-uint16 VNQ5050KTRE_LEFT_1STAT2_VAL;             //AIDN[3]
-uint16 VNQ5050KTRE_LEFT_1STAT3_VAL;             //ADIN[2]
-uint16 VNQ5050KTRE_LEFT_1STAT4_VAL;             //ADIN[10]
+uint8_t VNQ5050KTRE_LEFT_1STAT1_VAL;             //ADIN[11]
+uint8_t VNQ5050KTRE_LEFT_1STAT2_VAL;             //AIDN[3]
+uint8_t VNQ5050KTRE_LEFT_1STAT3_VAL;             //ADIN[2]
+uint8_t VNQ5050KTRE_LEFT_1STAT4_VAL;             //ADIN[10]
 
 /********************************************************4 channel POWER SWITCHE(RIGHT)***************************************************************/
 #define VNQ5050KTRE_RIGHT_INT1_PIN              1
@@ -66,10 +66,10 @@ uint16 VNQ5050KTRE_LEFT_1STAT4_VAL;             //ADIN[10]
 #define VNQ5050KTRE_RIGHT_STAT_DIS_PIN          18
 #define VNQ5050KTRE_RIGHT_STAT_DIS_PORT         hetPORT1
 
-uint16 VNQ5050KTRE_RIGHT_2STAT1_VAL;           //ADIN[8]
-uint16 VNQ5050KTRE_RIGHT_2STAT2_VAL;           //ADIN[6]
-uint16 VNQ5050KTRE_RIGHT_2STAT3_VAL;           //ADIN[5]
-uint16 VNQ5050KTRE_RIGHT_2STAT4_VAL;           //ADIN[4]
+uint8_t VNQ5050KTRE_RIGHT_2STAT1_VAL;           //ADIN[8]
+uint8_t VNQ5050KTRE_RIGHT_2STAT2_VAL;           //ADIN[6]
+uint8_t VNQ5050KTRE_RIGHT_2STAT3_VAL;           //ADIN[5]
+uint8_t VNQ5050KTRE_RIGHT_2STAT4_VAL;           //ADIN[4]
 
 /********************************************************2 channel POWER SWITCHE(LEFT)***************************************************************/
 #define VND5T100AJTRE_LEFT_INT1_PIN            2
@@ -81,8 +81,8 @@ uint16 VNQ5050KTRE_RIGHT_2STAT4_VAL;           //ADIN[4]
 #define VND5T100AJTRE_LEFT_FRSTBY_PIN          22           //PIN TO DO FAULT RESET
 #define VND5T100AJTRE_LEFT_FRSTBY_PORT         hetPORT1
 
-uint16 VNQ5050KTRE_LEFT_1CS1_VAL;             //ADIN[9]
-uint16 VNQ5050KTRE_LEFT_1CS2_VAL;             //ADIN[1]
+uint8_t VNQ5050KTRE_LEFT_1CS1_VAL;             //ADIN[9]
+uint8_t VNQ5050KTRE_LEFT_1CS2_VAL;             //ADIN[1]
 
 /********************************************************2 channel POWER SWITCHE(RIGHT)***************************************************************/
 #define VND5T100AJTRE_RIGHT_INT1_PIN            6
@@ -94,8 +94,8 @@ uint16 VNQ5050KTRE_LEFT_1CS2_VAL;             //ADIN[1]
 #define VND5T100AJTRE_RIGHT_FRSTBY_PIN          22           //PIN TO DO FAULT RESET
 #define VND5T100AJTRE_RIGHT_FRSTBY_PORT         hetPORT1
 
-uint16 VNQ5050KTRE_RIGHT_2CS1_VAL;             //ADIN[0]
-uint16 VNQ5050KTRE_RIGHT_2CS2_VAL;             //ADIN[7]
+uint8_t VNQ5050KTRE_RIGHT_2CS1_VAL;             //ADIN[0]
+uint8_t VNQ5050KTRE_RIGHT_2CS2_VAL;             //ADIN[7]
 
 /********************************************************1 channel POWER SWITCHE(BOTTOM CENTER)***************************************************************/
 #define VN5T006ASPTRE_BOTTOM_INT1_PIN            0
@@ -104,7 +104,7 @@ uint16 VNQ5050KTRE_RIGHT_2CS2_VAL;             //ADIN[7]
 #define VN5T006ASPTRE_BOTTOM_FRSTBY_PIN          22           //PIN TO DO FAULT RESET
 #define VN5T006ASPTRE_BOTTOM_FRSTBY_PORT         hetPORT1
 
-uint16 VN5T006ASPTRE_BOTTOM_CS1_VAL;            //ADIN[21]
+uint8_t VN5T006ASPTRE_BOTTOM_CS1_VAL;            //ADIN[21]
 
 /********************************************************1 channel POWER SWITCHE(BOTTOM RIGHT)***************************************************************/
 #define NCV8403ADTRKG_RIGHT_INT1_PIN            4
